speech_service/requirements.txt
faster-whisper
piper-tts
fastapi
uvicorn
prometheus-client
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-fastapi
